requirements.txt
falcon
falcon_cors
falcon-auth
PyYAML
python-ldap
regex
passlib
oyaml